DRY and move the WizardStep component to @syndesis/ui Move and simplify the IntegrationEditorChooseComponent to @syndesis/ui
