--- packages/ui/src/Integration/IntegrationEditorConnectionsList.tsx
@@ -1,8 +1,0 @@
-import { ListView } from 'patternfly-react';
-import * as React from 'react';
-
-export class IntegrationEditorConnectionsList extends React.Component<{}> {
-  public render() {
-    return <ListView>{this.props.children}</ListView>;
-  }
-}
