+++ packages/ui/src/Layout/WizardSteps.tsx
@@ -1,0 +1,32 @@
+import * as React from 'react';
+
+interface IWizardStepsProps {
+  mainSteps: React.ReactNode;
+  altSteps: React.ReactNode;
+  /**
+   * Indicates if the user clicked on a step. Used to show
+   * sub-steps when browsing from a mobile device.
+   */
+  active: boolean;
+}
+
+export class WizardSteps extends React.Component<IWizardStepsProps> {
+  public render() {
+    return (
+      <div className={'wizard-pf-steps'}>
+        <ul
+          className={`wizard-pf-steps-indicator wizard-pf-steps-alt-indicator ${
+            this.props.active ? 'active' : ''
+          }`}
+        >
+          {this.props.mainSteps}
+        </ul>
+        <ul
+          className={`wizard-pf-steps-alt ${this.props.active ? '' : 'hidden'}`}
+        >
+          {this.props.altSteps}
+        </ul>
+      </div>
+    );
+  }
+}
