+++ syndesis/src/modules/integrations/components/index.ts
@@ -1,6 +1,5 @@
 export * from './Integrations';
 export * from './IntegrationCreatorBreadcrumbs';
 export * from './IntegrationEditorBreadcrumbs';
-export * from './IntegrationEditorChooseConnection';
 export * from './IntegrationEditorSidebar';
 export * from './IntegrationEditorStepAdder';
